chore: rename auth & db server files
