+++ server/trpc/context.ts
@@ -3,8 +3,8 @@ import * as trpc from '@trpc/server'
 import * as trpcNext from '@trpc/server/adapters/next'
 import { Session } from 'next-auth'
 
-import { getServerAuthSession } from '../common/get-server-auth-session'
-import { prisma } from '../db/client'
+import { getServerAuthSession } from '../auth'
+import { prisma } from '../db'
 
 type CreateContextOptions = {
   session: Session | null
